In-Development: Preparing for OTAManager Network Request
Tried to get iOS-Common-Lib backported down to iOS 13... not easy. Shaders need to be skipped, which is doable, but Logger requires version 14.
