+++ Example/nRF Connect Device Manager.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -25,7 +25,7 @@
       "location" : "https://github.com/NordicSemiconductor/IOS-nRF-Connect-Device-Manager",
       "state" : {
         "branch" : "main",
-        "revision" : "7cf7fd71a7894b7edd5dca7b4c5f4b6af5f714f8"
+        "revision" : "fa03f3c66f8d81dd44b093dc5768b5e2bb2b08b3"
       }
     },
     {
